examples/python-ffmpeg/requirements.txt
ffmpeg-python
python-multipart
fastapi>=0.116.1
requests
